+++ ryan/src/java/org/apache/solr/handler/RequestHandlerBase.java
@@ -18,7 +18,6 @@
 package org.apache.solr.handler;
 
 import org.apache.solr.common.SolrException;
-import org.apache.solr.common.SolrDocumentList;
 import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.common.util.SimpleOrderedMap;
@@ -27,7 +26,6 @@ import org.apache.solr.core.SolrInfoMBean;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.request.SolrQueryResponse;
 import org.apache.solr.request.SolrRequestHandler;
-import org.apache.solr.search.DocSet;
 import org.apache.solr.util.SolrPluginUtils;
 import org.apache.lucene.queryParser.ParseException;
 
@@ -174,7 +172,7 @@ public abstract class RequestHandlerBase implements SolrRequestHandler, SolrInfo
     lst.add("timeouts", numTimeouts);
     lst.add("totalTime",totalTime);
     lst.add("avgTimePerRequest", (float) totalTime / (float) this.numRequests);
-    lst.add("avgRequestsPerSecond", (float) numRequests*1000 / (float)(System.currentTimeMillis()-handlerStart));   
+    lst.add("avgRequestsPerSecond", (float) numRequests*1000 / (System.currentTimeMillis()-handlerStart));   
     return lst;
   }
